test(webdav): separate delete header unit test
